# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Media Streaming Platform API
  version: 1.0.0
  description: Public demo API. Accounts and viewing data may be reset.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /movies/{slug}:
    get:
      operationId: movieBySlug
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Movie details
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MovieDetails'}
        '404': {$ref: '#/components/responses/Problem'}
...
  /movies/{movieId}/playback:
    get:
      operationId: playback
      parameters:
        - {$ref: '#/components/parameters/MovieId'}
      responses:
        '200':
          description: >-
            Managed movies answer a version-pinned grant with a short-lived media token; legacy
            fixtures keep the manifest/resume response and omit the managed fields.
          content:
            application/json:
              schema:
                anyOf:
                  - {$ref: '#/components/schemas/ManagedPlayback'}
                  - {$ref: '#/components/schemas/Playback'}
        '401': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
...
  /me/progress/{movieId}:
    put:
      operationId: updateProgress
      parameters:
        - {$ref: '#/components/parameters/MovieId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProgressUpdate'}
      responses:
        '200':
          description: Stored progress
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Progress'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /admin/movies:
    get:
      operationId: adminListMovies
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      responses:
        '200':
          description: Managed and legacy movie page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMoviePage'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
    post:
      operationId: adminCreateMovie
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, maxLength: 200}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/MovieInput'}
      responses:
        '201':
          description: Created draft
          headers:
            ETag: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMovie'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /admin/movies/{movieId}:
    get:
      operationId: adminGetMovie
      parameters:
        - {$ref: '#/components/parameters/MovieId'}
      responses:
        '200':
          description: Movie with revision
          headers:
            ETag: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMovie'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
    put:
      operationId: adminUpdateMovie
      parameters:
        - {$ref: '#/components/parameters/MovieId'}
        - name: If-Match
          in: header
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/MovieInput'}
      responses:
        '200':
          description: Updated movie
          headers:
            ETag: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMovie'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '412': {$ref: '#/components/responses/Problem'}
        '428': {$ref: '#/components/responses/Problem'}
  /admin/genres:
    get:
      operationId: adminListGenres
      responses:
        '200':
          description: Available genres
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/GenreOption'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
...
  /admin/movies/{movieId}/artwork:
    post:
      operationId: adminAttachArtwork
      parameters:
        - {$ref: '#/components/parameters/MovieId'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ArtworkRequest'}
      responses:
        '200':
          description: Movie with the selected artwork pointer
          headers:
            ETag: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMovie'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '412': {$ref: '#/components/responses/Problem'}
        '428': {$ref: '#/components/responses/Problem'}
...
  /admin/movies/{movieId}/publish:
    post:
      operationId: adminPublishMovie
      parameters:
        - {$ref: '#/components/parameters/MovieId'}
        - {$ref: '#/components/parameters/IfMatch'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PublishRequest'}
      responses:
        '200':
          description: Published movie with the active version and artwork pointers
          headers:
            ETag: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMovie'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '412': {$ref: '#/components/responses/Problem'}
        '428': {$ref: '#/components/responses/Problem'}
  /admin/movies/{movieId}/activate:
    post:
      operationId: adminActivateVersion
      parameters:
        - {$ref: '#/components/parameters/MovieId'}
        - {$ref: '#/components/parameters/IfMatch'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PublishRequest'}
      responses:
        '200':
          description: Movie with the newly activated version
          headers:
            ETag: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMovie'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '412': {$ref: '#/components/responses/Problem'}
        '428': {$ref: '#/components/responses/Problem'}
  /admin/movies/{movieId}/unpublish:
    post:
      operationId: adminUnpublishMovie
      parameters:
        - {$ref: '#/components/parameters/MovieId'}
        - {$ref: '#/components/parameters/IfMatch'}
      responses:
        '200':
          description: UNPUBLISHED movie; the active version is retained
          headers:
            ETag: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMovie'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '412': {$ref: '#/components/responses/Problem'}
        '428': {$ref: '#/components/responses/Problem'}
  /admin/movies/{movieId}/archive:
    post:
      operationId: adminArchiveMovie
      parameters:
        - {$ref: '#/components/parameters/MovieId'}
        - {$ref: '#/components/parameters/IfMatch'}
      responses:
        '200':
          description: ARCHIVED movie
          headers:
            ETag: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMovie'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '412': {$ref: '#/components/responses/Problem'}
        '428': {$ref: '#/components/responses/Problem'}
  /admin/movies/{movieId}/restore:
    post:
      operationId: adminRestoreMovie
      parameters:
        - {$ref: '#/components/parameters/MovieId'}
        - {$ref: '#/components/parameters/IfMatch'}
      responses:
        '200':
          description: Restored UNPUBLISHED movie; restore never republishes
          headers:
            ETag: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMovie'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '412': {$ref: '#/components/responses/Problem'}
        '428': {$ref: '#/components/responses/Problem'}
  /admin/movies/{movieId}/preview:
    post:
      operationId: adminPreviewMovie
      parameters:
        - {$ref: '#/components/parameters/MovieId'}
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PublishRequest'}
      responses:
        '200':
          description: >-
            Short-lived ADMIN preview grant for the requested READY version, or for the active or
            latest READY version when no mediaVersionId is supplied
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ManagedPlayback'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Size:
      name: size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    MovieId:
      name: movieId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, maxLength: 200}
  responses:
...
    Problem:
      description: RFC 9457 Problem Details with stable extensions
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
...
    MovieSummary:
      type: object
      required: [id, slug, title, releaseYear, runtimeSeconds, maturityRating, posterUrl, backdropUrl, genres]
      properties:
        id: {type: string, format: uuid}
        slug: {type: string}
        title: {type: string}
        releaseYear: {type: integer}
        runtimeSeconds: {type: integer}
        maturityRating: {type: string}
        posterUrl: {type: string, format: uri-reference}
        backdropUrl: {type: string, format: uri-reference}
        genres: {type: array, items: {type: string}}
    MovieDetails:
      allOf:
        - {$ref: '#/components/schemas/MovieSummary'}
        - type: object
          required: [synopsis, playable]
          properties:
            synopsis: {type: string}
            playable: {type: boolean}
    MovieInput:
      type: object
      additionalProperties: false
      required: [title, slug, releaseYear, maturityRating, featured]
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        slug: {type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$', maxLength: 120}
        synopsis: {type: string, maxLength: 10000}
        releaseYear: {type: integer, minimum: 1888, maximum: 2200}
        maturityRating: {type: string, enum: [G, PG, PG-13, R, NC-17, NR]}
        genreIds:
          type: array
          uniqueItems: true
          maxItems: 10
          items: {type: integer, minimum: 1}
        featured: {type: boolean}
    AdminMovie:
      type: object
      required: [id, title, slug, synopsis, releaseYear, maturityRating, genreIds, featured,
                 lifecycle, revision, managementMode, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        title: {type: string}
        slug: {type: string}
        synopsis: {type: string}
        releaseYear: {type: integer}
        maturityRating: {type: string}
        genreIds: {type: array, items: {type: integer}}
        featured: {type: boolean}
        lifecycle: {type: string, enum: [DRAFT, PUBLISHED, UNPUBLISHED, ARCHIVED]}
        revision: {type: integer, format: int64}
        managementMode: {type: string, enum: [LEGACY, MANAGED]}
        activeMediaVersionId: {type: [string, 'null'], format: uuid}
        posterAssetId: {type: [string, 'null'], format: uuid}
        backdropAssetId: {type: [string, 'null'], format: uuid}
        runtimeSeconds: {type: [integer, 'null']}
        firstPublishedAt: {type: [string, 'null'], format: date-time}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    AdminMoviePage:
      type: object
      required: [items, page, size, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/AdminMovie'}}
        page: {type: integer}
        size: {type: integer}
        total: {type: integer, format: int64}
    GenreOption:
      type: object
      required: [id, slug, name]
      properties:
        id: {type: integer}
        slug: {type: string}
        name: {type: string}
...
    Playback:
      type: object
      required: [movieId, manifestUrl, resumePositionSeconds]
      properties:
        movieId: {type: string, format: uuid}
        manifestUrl: {type: string, format: uri-reference}
        resumePositionSeconds: {type: integer, minimum: 0}
    ManagedPlayback:
      type: object
      required: [movieId, manifestUrl, resumePositionSeconds, sessionId, mediaVersionId, mediaToken,
                 mediaTokenExpiresAt]
      properties:
        movieId: {type: string, format: uuid}
        manifestUrl: {type: string, format: uri-reference}
        resumePositionSeconds: {type: integer, minimum: 0}
        sessionId: {type: string, format: uuid}
        mediaVersionId: {type: string, format: uuid}
        mediaToken: {type: string}
        mediaTokenExpiresAt: {type: string, format: date-time}
...
    PublishRequest:
      type: object
      additionalProperties: false
      required: [mediaVersionId]
      properties:
        mediaVersionId: {type: string, format: uuid}
        posterAssetId: {type: [string, 'null'], format: uuid}
        backdropAssetId: {type: [string, 'null'], format: uuid}
    ArtworkRequest:
      type: object
      additionalProperties: false
      required: [kind, assetId]
      properties:
        kind: {type: string, enum: [POSTER, BACKDROP]}
        assetId: {type: string, format: uuid}
...
    ProgressUpdate:
      type: object
      additionalProperties: false
      required: [positionSeconds, durationSeconds, clientUpdatedAt]
      properties:
        positionSeconds: {type: integer, minimum: 0}
        durationSeconds: {type: integer, minimum: 1}
        clientUpdatedAt: {type: string, format: date-time}
        sessionId: {type: string, format: uuid}
        mediaVersionId: {type: string, format: uuid}
    Progress:
      allOf:
        - {$ref: '#/components/schemas/ProgressUpdate'}
        - type: object
          required: [movieId, completed]
          properties:
            movieId: {type: string, format: uuid}
            completed: {type: boolean}
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: {type: string}
        code: {type: string}
        message: {type: string}
    Problem:
      type: object
      required: [type, title, status, detail, requestId, code]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        requestId: {type: string}
        code: {type: string}
        fieldErrors:
          type: array
          items: {$ref: '#/components/schemas/FieldError'}